BUGFIX: prevent infinit loop due to doundoff error
use 10e-8 (not 10e-16) to be conservative.

for flat target: better keep track if target was the valid hit during
the last round. if yes, the possible hit in this round is just the
starting point of the ray fuzzed by round off errors. in that case
return NULL (no hit)

Signed-off-by: Ivo Alxneit <alxneit@rutil.(none)>