BUGFIX: only mark reflected rays with 'las_was_hit'
absorbed rays are terminated. if they are marked 'last_was_hit', as was
done before, the next in_ray will falsely be declared as originating
on this target and 'get_intercept()' will falsely return NULL. only
non-absorbed rays need to be marked as they continue in the next round.

Signed-off-by: Ivo Alxneit <alxneit@rutil.(none)>

Conflicts:

	target_square.c
	target_triangle.c