pass 'r' to 'xxx_get_out_ray()' and do check for absorption there.
this is cleaner implementation. it should be somewhat faster as the test
for ray_absorbed_due_to_non-unit_reflectivity is performed only on the
nearest target (the one from which the ray continues) and not one every
target positioned further away on the path of the ray. on each we save
a call to 'gsl_rng_uniform()' and a spline interpolation ('gsl_spline_eval()').

Signed-off-by: Ivo Alxneit <ivo.alxneit@psi.ch>