BUGFIX: wrong header / comments for transformation local <-> global
in all comments and also in the header of the output files the wrong
transformation was given. this was very confusing as the header of the
output stated that M is applied
	g(x, y, z) = M l(x, y, z) + origin(x, y, z)
while MT (transpose) has to be applied
	g(x, y, z) = MT l(x, y, z) + origin(x, y, z)

nothing was changed in the code (was correct) but now header, comments, and
code are consistent.

Signed-off-by: Ivo Alxneit <ivo.alxneit@psi.ch>