keep track of number of reflections each ray experiences.
ray->n_refl will wrap around at 256 reflections but this limit will only
be reached in pathological cases (or if "many" targets are involved such
as when e.g. calculating the transmission of RPC's). an unsigned 8bit
integer seems a good compromise between increasing the output files and
implementing a safe maximum.

Signed-off-by: Ivo Alxneit <alxneit@rutil>