move reading data into binned()
this is a preparatory step to reading and processing data in chunks to
put less strain on the memory.
for 'cylinder' and 'sphere' we read to first line in their specific 'bin_X()'
function as we need the radius to calculated the area segments. thus we
loose one data set in the binning process. as we typically bin a lot of
data, this is no problem

Signed-off-by: Ivo Alxneit <ivo.alxneit@psi.ch>